Revert "merge with 3.18.44"
This reverts commit b35813598b444852da9b53c9732997f8ef3e555f, reversing
changes made to f661eb35f108ed5023332ba6425b7e8eae9ecb1c.

# Conflicts:
#	mm/page_alloc.c

+++ arch/sparc/kernel/sys_sparc_64.c
@@ -413,7 +413,7 @@ out:
 
 SYSCALL_DEFINE1(sparc64_personality, unsigned long, personality)
 {
-	long ret;
+	int ret;
 
 	if (personality(current->personality) == PER_LINUX32 &&
 	    personality(personality) == PER_LINUX)
